pyproject.toml
[tool.poetry]
name = "raise-hackathon-prosus-track"
version = "0.1.0"
description = ""
authors = ["Ashwath Shetty <64685993+Ashwath-Shetty@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "raise_hackathon_prosus_track"}]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.3.0"
jupyterlab = "^4.4.4"
groq = "0.29.0"
gradio = "5.35.0"
httpx = "0.28.1"
langchain = "0.3.26"
langchain-groq = "0.3.5"
matplotlib = "3.10.3"
networkx = "3.4.2"
pillow = "11.3.0"
pydantic = "2.11.7"
requests = "2.32.4"
python-dotenv = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
